+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 ### Version 7.8
 - Edit ingredient category rewrite
+- CSS update adding bs4 styling
+- Update tagsinput to bs4 compatibility
 
 ### Version 7.7
 - Fix a bug incorrectly showing available solvents when adjust was enabled in ingredient add