+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # CHANGELOG
+### Version 7.9
 ### Version 7.8
 - Fix a bug causing ingredients synonym search to fail
 - Edit ingredient category rewrite